requirements.txt
rev-ai
pyaudio
python-dotenv
scipy
pyannote.audio
torch
colorama